# openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "description": "Returns a list of all registered users.",
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "Users Management"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retrieves user details and the list of cats they have adopted.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found."
          },
          "400": {
            "description": "Invalid ID format provided."
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "Users Management"
        ]
      },
      "delete": {
        "description": "Permanently removes a user. Note: Associated cats will remain in the system but will become ownerless (SET NULL).",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID to delete",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully."
          },
          "400": {
            "description": "Invalid ID"
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "403": {
            "description": "Access denied. You can only delete your own profile"
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete user profile",
        "tags": [
          "Users Management"
        ]
      }
    },
    "/users/{id}/cats": {
      "get": {
        "description": "Returns user profile with the list of their adopted cats.",
        "operationId": "UsersController_getUserCats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID"
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all cats adopted by a specific user",
        "tags": [
          "Users Management"
        ]
      }
    },
    "/users/{id}/make-admin": {
      "post": {
        "description": "Changes the user role to Admin. Restricted to users with existing Admin privileges.",
        "operationId": "UsersController_makeAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID to be promoted",
            "schema": {
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User has been successfully promoted to Admin."
          },
          "400": {
            "description": "Invalid ID format provided."
          },
          "401": {
            "description": "Unauthorized: Token is missing or invalid."
          },
          "403": {
            "description": "Forbidden: You do not have administrator rights."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Give admin rights to user",
        "tags": [
          "Users Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Приют для кошек",
    "description": "Система управления кошками и пользователями",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserEntity": {
        "type": "object",
        "properties": {

        }
      },
...
    }
  }
}